include base packages used by required python modules
The requirements.txt for Automation Tools includes a number of dependencies
that need building, such as lxml and sqlalchemy. These need build tools and
libs installed for the build to succeed.

In the case of lxml we actually install this as part of the Dockerfile, as
it takes quite a long time and so we don't want to have to do it for each
instantiation of the image.